KikiPos: accept float params in python3
The whole "py" dir uses code with simple "/" operator which does not
always returns an int even when its operand are ints, like it did in
python2.

Examining every "/" occurence to determine if we should convert to int
after the computation would take a long time. Instead, just accept
floats as KikiPos parameters, converting to int on the C++ side.